+++ Stores/SqlServer/Cleipnir.ResilientFunctions.SqlServer/SqlGenerator.cs
@@ -394,7 +394,7 @@ public class SqlGenerator(string tablePrefix)
     }
     
     private string? _getMessagesSql;
-    public StoreCommand GetMessages(StoredId storedId, int skip, string paramPrefix = "")
+    public StoreCommand GetMessages(StoredId storedId, long skip, string paramPrefix = "")
     {
         _getMessagesSql ??= @$"
             SELECT Content
@@ -436,16 +436,16 @@ public class SqlGenerator(string tablePrefix)
     
     public async Task<Dictionary<StoredId, List<byte[]>>> ReadStoredIdsMessages(SqlDataReader reader)
     {
-        var messages = new Dictionary<StoredId, List<Tuple<int, byte[]>>>();
+        var messages = new Dictionary<StoredId, List<Tuple<long, byte[]>>>();
 
         while (await reader.ReadAsync())
         {
             var id = reader.GetGuid(0).ToStoredId();
-            var position = reader.GetInt32(1);
+            var position = reader.GetInt64(1);
             var content = (byte[]) reader.GetValue(2);
 
             if (!messages.ContainsKey(id))
-                messages[id] = new List<Tuple<int, byte[]>>();
+                messages[id] = new List<Tuple<long, byte[]>>();
 
             messages[id].Add(Tuple.Create(position, content));
         }
